Make the diff the primary link in thanks
When viewing a thank message, make the link to the diff the primary link.
Visiting the user's user page is almost never the desired action when
viewing a thank notification. The link to the user page is no longer
present as a secondary link, but the notification-thanks-flyout2 and
notification-thanks messages both include a link from the user's name to
their user page.

Bug: 56225
Change-Id: If971385e6d1a47e394d06decf25db2c1b7cc2fb9