Avoid DatabaseBlock address parameter
The address parameter will soon be deprecated, so don't use it.

There is a factory for DatabaseBlock objects now which should be used in
integration tests. Since this is pretending to be a unit test, I will
demonstrate the new parameters for direct construction of DatabaseBlock
objects.

I don't think it's OK to construct a DatabaseBlock in a provider, since
DatabaseBlock still requires service container access.

Bug: T385966
Change-Id: I18e314012b75f183ae312f4259eb6a353cab4c0b