Fix Echo test failures when Flow is not loaded
Follow-up to 9f58bd28a7a6be99a3c79d17ac1165ab5824033c.

I thought that the Flow thank event can be registered unconditionally,
but there is actually a test in Echo (NotificationStructureTest)
that fails when the specified presentation model doesn't exist.

Restore the hook to conditionally unregister it (which is actually
more in line with what Echo itself does with optional events).

Bug: T394565
Change-Id: I6917a2893fd20d198e16845415dc24179bc30f28