Fix client-side duplicate thanks prevention
Why:
- Session keys used for checking if the user has already
  thanked for a revision were inconsistent.
- ApiCoreThank::sendThanks set "thanks-thanked-rev#" (since
  ed087b0) for revisions, but in ::userAlreadySentThanks it
  checked "thanks-thanked-#", and the same check was done in
  Hooks::generateThankElement.

What:
- Centralize code where the session key is created. This reveals
  another inconsistency between the interfaces in literals
  indicating it's a revision what is being thanked for. The new
  method makes sure the same key is produced.

Note that this patch drops code claiming to maintain backward
compatibility. It was supposed to maintain compatibility with
keys stored prior to ed087b0, but the author forgot to check
the new keys (...-rev#), too. Thus, we update the code to be
compatible with the keys produced for the last 7 years.

Change-Id: Ibfa065563d8f77738f4277cd947acd425bc3b4d9